requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.7
django-better-admin-arrayfield==1.3.0
django-crispy-forms==1.9.2
django-foundation-filefield-widget==0.0.4
django-json-widget==1.0.1
djangorestframework==3.11.1
furl==2.1.0
future==0.18.2
idna==2.10
numpy==1.19.2
orderedmultidict==1.0.1
pandas==1.1.2
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
wincertstore==0.2
xmltodict==0.12.0
